Move ModelParams code into fwdpy11._types
